Add files via upload
On this PART A, you are required to build a command prompt (console) application.The applicationmust ask from the command prompt to input data to allthe entitiesand it should give the option to add more than one entry at a time[10 marks].If the user decides not to type any data,the program should use synthetic data[5 marks].Also, your program must be able to output the following:•A list of all the students[5 marks]•A list of all the trainers[5 marks]•A list of all the assignments[5 marks]•A list of all the courses[5 marks]•All the students per course[10 marks]HINT:Make a new class that contains the students within a course•All the trainers per course[10 marks]HINT: As above•All the assignments per course[10 marks]HINT: As above•All the assignments per student[10 marks]HINT:You need to relate a student per course and per assignment•A list of students that belong to more than one courses[10 marks]Lastly, the program should ask from the user a date and it should output a list of students who need to submit one or more assignments on the same calendar week as thegiven date[15marks].
HINT1:The calendar week is consideredas Monday to Fridayfor submissions. For example,if the user inputs 15/2/2019you need to check from Monday 11/2/2019 toFriday15/2/2019.If the userinputs 16/2/2019(Saturday) or 17/2/2019 (Sunday)you need to check from Monday 11/2/2019 to Friday 15/2/2019.